Give app icon designer full control over icon use.
Fixed a flaw in the terms of icon use which may lead others to believe that we have full control over the app icon. Not really. We only removed the background to make it look like a favicon. The icon was the designer's own work. Therefore, please contact the icon designer for use rights.